+++ lib/index.js
@@ -1,4 +1,5 @@
 const prompts = require("./prompts")
 const queries = require("./queries")
+const handlers = require("./handlers")
 
-module.exports = { prompts, queries }
+module.exports = { prompts, queries, handlers }
